arma-support-web/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.32
PyMySQL==1.1.1
python-dotenv==1.0.1
jinja2==3.1.4
python-multipart==0.0.9
itsdangerous==2.2.0
passlib[bcrypt]==1.7.4
